feat: FMatrix 분리
